+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Release 0.X.0 (unreleased)
 * Update to CycloneDX spec 1.4 (#296)
 * Never overwrite main project folder and manifest (#302)
 * Add codespell and fix typo's (#303)
+* Add warning to metadata file, not to change it (#170)
 
 Release 0.6.0 (released 2022-01-31)
 ===================================